refactor: split functionality into files
