+++ package/src/swagger_server/controllers/db.py
@@ -1,0 +1,5 @@
+from pymongo import MongoClient
+
+CLIENT = MongoClient('mongodb://mongodb:27017/')
+DB = CLIENT.database
+query_details = DB.query_db
