+++ package/src/swagger_server/controllers/helpers.py
@@ -1,5 +1,6 @@
 from bson import ObjectId
 
+
 def check_id(f):
 
     def wrapper(*args, **kwargs):
@@ -10,3 +11,10 @@ def check_id(f):
         return f(*args, **kwargs)
 
     return wrapper
+
+def _stringify_object_id(result):
+    stringified_result = []
+    for element in result:
+        element['_id'] = str(element['_id'])
+        stringified_result.append(element)
+    return stringified_result
